+++ tpu/src/test/java/tpu/TpuVmIT.java
@@ -17,6 +17,7 @@
 package tpu;
 
 import static com.google.common.truth.Truth.assertThat;
+import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
@@ -25,6 +26,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.google.api.gax.longrunning.OperationFuture;
+import com.google.cloud.tpu.v2.CreateNodeRequest;
 import com.google.cloud.tpu.v2.DeleteNodeRequest;
 import com.google.cloud.tpu.v2.GetNodeRequest;
 import com.google.cloud.tpu.v2.Node;
@@ -42,11 +44,13 @@ import org.junit.runners.JUnit4;
 import org.mockito.MockedStatic;
 
 @RunWith(JUnit4.class)
-@Timeout(value = 3)
+@Timeout(value = 30)
 public class TpuVmIT {
   private static final String PROJECT_ID = "project-id";
   private static final String ZONE = "asia-east1-c";
   private static final String NODE_NAME = "test-tpu";
+  private static final String TPU_TYPE = "v2-8";
+  private static final String TPU_SOFTWARE_VERSION = "tpu-vm-tf-2.12.1";
   private static ByteArrayOutputStream bout;
 
   @BeforeAll
@@ -56,19 +60,42 @@ public class TpuVmIT {
   }
 
   @Test
+  public void testCreateTpuVm() throws Exception {
+    try (MockedStatic<TpuClient> mockedTpuClient = mockStatic(TpuClient.class)) {
+      Node mockNode = mock(Node.class);
+      TpuClient mockTpuClient = mock(TpuClient.class);
+      OperationFuture mockFuture = mock(OperationFuture.class);
+
+      mockedTpuClient.when(() -> TpuClient.create(any(TpuSettings.class)))
+          .thenReturn(mockTpuClient);
+      when(mockTpuClient.createNodeAsync(any(CreateNodeRequest.class)))
+          .thenReturn(mockFuture);
+      when(mockFuture.get()).thenReturn(mockNode);
+
+      Node returnedNode = CreateTpuVm.createTpuVm(
+          PROJECT_ID, ZONE, NODE_NAME,
+          TPU_TYPE, TPU_SOFTWARE_VERSION);
+
+      verify(mockTpuClient, times(1))
+          .createNodeAsync(any(CreateNodeRequest.class));
+      verify(mockFuture, times(1)).get();
+      assertEquals(returnedNode, mockNode);
+    }
+  }
+
+  @Test
   public void testGetTpuVm() throws IOException {
     try (MockedStatic<TpuClient> mockedTpuClient = mockStatic(TpuClient.class)) {
       Node mockNode = mock(Node.class);
       TpuClient mockClient = mock(TpuClient.class);
-      GetTpuVm mockGetTpuVm = mock(GetTpuVm.class);
 
       mockedTpuClient.when(TpuClient::create).thenReturn(mockClient);
       when(mockClient.getNode(any(GetNodeRequest.class))).thenReturn(mockNode);
 
       Node returnedNode = GetTpuVm.getTpuVm(PROJECT_ID, ZONE, NODE_NAME);
 
-      verify(mockGetTpuVm, times(1))
-          .getTpuVm(PROJECT_ID, ZONE, NODE_NAME);
+      verify(mockClient, times(1))
+          .getNode(any(GetNodeRequest.class));
       assertThat(returnedNode).isEqualTo(mockNode);
     }
   }
